SFR-369 Refactor Search class
Move the `Search` class into its own file in the `lib` directory
for better separation of concerns and ease of maintainability.

Refactor a few associated files and remove the unused `client.js`
file.
