--- lib/client.js
@@ -1,34 +1,0 @@
-const Elastic = require('elasticsearch')
-require('dotenv').config()
-
-function elasticsearchClient () {
-  const client = new Elastic.Client({
-    host: process.env.ELASTICSEARCH_HOST,
-    log: 'trace'
-  })
-
-  this.healthCheck = function () {
-    client.ping({
-      requestTimeout: 30000
-    }, function (error) {
-      if (error) {
-        console.error('Cluster is unreachable')
-      } else {
-        console.log('Healthy instance!')
-      }
-    })
-  }
-
-  this.simpleSearch = function () {
-    client.search({
-      q: 'gutenberg'
-    }).then(function (body) {
-      const hits = body.hits.hits
-      console.log(hits)
-    }, function (error) {
-      console.trace(error)
-    })
-  }
-}
-
-module.exports = { elasticsearchClient }
